requirements.txt
ultralytics>=8.2.0
PyYAML>=6.0.1
opencv-python>=4.8.0
